tcos_quiz_bot/config.py

import os
from dotenv import load_dotenv

# Load variables from .env file
load_dotenv()

# Telegram Bot Token
BOT_TOKEN = os.getenv("TELEGRAM_BOT_TOKEN")

# Optional: Admin Chat ID for notifications
ADMIN_CHAT_ID = os.getenv("ADMIN_CHAT_ID")

if not BOT_TOKEN:
    raise ValueError("No TELEGRAM_BOT_TOKEN found in environment variables. Please check your .env file.")
