tcos_quiz_bot/requirements.txt

python-telegram-bot==20.8
python-dotenv==1.0.1
Pillow==10.2.0
